a separate method for getting different kinds of content; storing context as case class attachments---diff for each type of reader, but can be easily printed regardless of differences in context contents
