+++ src/main/scala/org/clulab/habitus/utils/JsonlPrinter.scala
@@ -4,10 +4,10 @@ import org.clulab.odin.Mention
 import org.clulab.processors.Document
 import org.clulab.serialization.json.stringify
 import org.clulab.utils.FileUtils
+import org.json4s.JsonAST.JObject
 import org.json4s.JsonDSL._
 
 import java.io.PrintWriter
-import scala.collection.mutable
 
 class JsonlPrinter(outputFilename: String) extends Printer {
   protected val printWriter: PrintWriter = FileUtils.printWriterFromFile(outputFilename)
@@ -19,7 +19,6 @@ class JsonlPrinter(outputFilename: String) extends Printer {
   protected def outputMention(
     mention: Mention,
     doc: Document,
-    contextDetailsMap: mutable.Map[Int, ContextDetails],
     inputFilename: String,
     printVars: PrintVariables
   ): Unit = {
@@ -35,47 +34,21 @@ class JsonlPrinter(outputFilename: String) extends Printer {
           valueMention.lemmas
               .map(_.mkString(" "))
               .getOrElse(valueMention.text)
-    val jObject = contextDetailsMap
-        .get(mention.sentence) // Get it, optionally, if it is there.
-        .map { contextDetails => // If there, do this.
-          ("variableText" -> variableText) ~
-          ("valueText" -> valueText) ~
-          ("valueNorm" -> valueNorm) ~
-          ("sentenceText" -> sentenceText) ~
-          ("inputFilename" -> inputFilename) ~
-          ("mostFreqLoc0Sent" -> contextDetails.mostFreqLoc0Sent) ~
-          ("mostFreqLoc1Sent" -> contextDetails.mostFreqLoc1Sent) ~
-          ("mostFreqLoc" -> contextDetails.mostFreqLoc) ~
-          ("mostFreqDate0Sent" -> contextDetails.mostFreqDate0Sent) ~
-          ("mostFreqDate1Sent" -> contextDetails.mostFreqDate1Sent) ~
-          ("mostFreqDate" -> contextDetails.mostFreqDate) ~
-          ("mostFreqCrop0Sent" -> contextDetails.mostFreqCrop0Sent) ~
-          ("mostFreqCrop1Sent" -> contextDetails.mostFreqCrop1Sent) ~
-          ("mostFreqCrop" -> contextDetails.mostFreqCrop) ~
-          ("mostFreqFert0Sent" -> contextDetails.mostFreqFertilizer0Sent) ~
-          ("mostFreqFert1Sent" -> contextDetails.mostFreqFertilizer1Sent) ~
-          ("mostFreqFert" -> contextDetails.mostFreqFertilizerOverall)
-        }
-        .getOrElse { // If it wasn't there, do this instead.
-          // These keys should match the ones used above.
-          ("variableText" -> variableText) ~
-            ("valueText" -> valueText) ~
-            ("valueNorm" -> valueNorm) ~
-            ("sentenceText" -> sentenceText) ~
-            ("inputFilename" -> inputFilename) ~
-            ("mostFreqLoc0Sent" -> "N/A") ~
-            ("mostFreqLoc1Sent" -> "N/A") ~
-            ("mostFreqLoc" -> "N/A") ~
-            ("mostFreqDate0Sent" -> "N/A") ~
-            ("mostFreqDate1Sent" -> "N/A") ~
-            ("mostFreqDate" -> "N/A") ~
-            ("mostFreqCrop0Sent" -> "N/A") ~
-            ("mostFreqCrop1Sent" -> "N/A") ~
-            ("mostFreqCrop" -> "N/A") ~
-            ("mostFreqFert0Sent" -> "N/A") ~
-            ("mostFreqFert1Sent" -> "N/A") ~
-            ("mostFreqFert" -> "N/A")
-        }
+
+    var jObject: JObject =
+      ("variableText" -> variableText) ~
+      ("valueText" -> valueText) ~
+      ("valueNorm" -> valueNorm) ~
+      ("sentenceText" -> sentenceText) ~
+      ("inputFilename" -> inputFilename)
+
+    // add key-value pairs from the context attachment
+    mention.attachments.head.asInstanceOf[Context].getArgs()
+      .foreach {
+        i =>
+          jObject = jObject ~ (i._1 -> i._2.toString)
+      }
+
     val json = stringify(jObject, pretty = false)
     val jsonl = json.replace('\n', ' ') // just in case
 
@@ -85,17 +58,16 @@ class JsonlPrinter(outputFilename: String) extends Printer {
   def outputMentions(
     mentions: Seq[Mention],
     doc: Document,
-    contextDetailsMap: mutable.Map[Int, ContextDetails],
     inputFilename: String,
     printVars: PrintVariables
   ): Unit = {
     println(s"Writing mentions from doc $inputFilename to $outputFilename")
     mentions
         .filter { mention => mention.label.matches(printVars.mentionLabel) }
-        .filter { mention => contextDetailsMap.isEmpty || contextDetailsMap.contains(mention.sentence) }
+//        .filter { mention => contextDetailsMap.isEmpty || contextDetailsMap.contains(mention.sentence) }
         .sortBy { mention => (mention.sentence, mention.start) }
         .foreach { mention =>
-          outputMention(mention, doc, contextDetailsMap, inputFilename, printVars)
+          outputMention(mention, doc, inputFilename, printVars)
         }
     printWriter.flush()
   }
